pyproject.toml
[tool.poetry]
name = "study-materials-generator-pkg"
version = "0.1.0"
description = "This project aims to develop an intelligent study assistant application designed to empower high school students in their learning process.  Leveraging the power of AI, this application will transform static study materials, such as notes and quizzes (in PNG and PDF formats), into dynamic and interactive learning experiences."
authors = ["Bodie <46969107+riddles-in-the-dark@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytesseract = "^0.3.13"
pillow = "^11.1.0"
pypdf2 = "^3.0.1"
sentence-transformers = "^3.4.1"
requests = "^2.32.3"
aiohttp = "^3.11.12"
ollama = "^0.4.7"
scikit-learn = "^1.6.1"
transformers = "^4.49.0"
nest-asyncio = "^1.6.0"
bitsandbytes = "^0.39.0"
accelerate = ">=0.26.0"
streamlit = "1.42.2"
altair = "5.5.0"
annotated-types = "0.7.0"
anyio = "4.8.0"
asyncio = "3.4.3"
attrs = "25.1.0"
blinker = "1.9.0"
cachetools = "5.5.2"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
colorama = "0.4.6"
filelock = "3.17.0"
fsspec = "2025.2.0"
gitdb = "4.0.12"
GitPython = "3.1.44"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
huggingface-hub = "0.29.1"
idna = "3.10"
Jinja2 = "3.1.5"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
markdown-it-py = "3.0.0"
MarkupSafe = "3.0.2"
mdurl = "0.1.2"
mpmath = "1.3.0"
narwhals = "1.27.1"
networkx = "3.4.2"
numpy = "2.2.3"
packaging = "24.2"
pandas = "2.2.3"
protobuf = "5.29.3"
pyarrow = "19.0.1"
pydantic = "2.10.6"
pydantic_core = "2.27.2"
pydeck = "0.9.1"
Pygments = "2.19.1"
python-dateutil = "2.9.0.post0"
pytz = "2025.1"
PyYAML = "6.0.2"
referencing = "0.36.2"
regex = "2024.11.6"
rich = "13.9.4"
rpds-py = "0.23.1"
safetensors = "0.5.2"
scipy = "1.15.2"
setuptools = "75.8.0"
six = "1.17.0"
smmap = "5.0.2"
sniffio = "1.3.1"
sympy = "1.13.1"
tenacity = "9.0.0"
threadpoolctl = "3.5.0"
tokenizers = "0.21.0"
toml = "0.10.2"
torch = "2.6.0"
tornado = "6.4.2"
tqdm = "4.67.1"
typing_extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
